wip, first pass at a 'Idle switch' plugin
If there is no input events on an 'input source'
channel for some period of time, flip a bit high.

eventually outs should be a gate?  maybe a
idle start/end triggers?

time period is just hooked up to know widget at
the moment, but will eventually be based on
a clock like heartbeat input.