Stop clobbering ModuleResizeHandle widget from app
Not sure how, but with this using the same name as
the resize handler in core/Blank.hpp, my instances
were using that widget instead of the local one.

Only noticible diff was that I had removed the
handle line/shadows.

The behavior in this case seemed to be different
between v0.5 and v0.6. But that could have to
with runtime ordering I suppose...

I was assuming I couldnt/wouldn't clobber the name
of Core widgets because they werent obviously in
the include, but not sure.

Is something finding the widgets by name somewhere?
(vaguely recall something about name base resolving
but no idea where...)

Something related to widgets.cpp gHoveredWidget and
friends? I'm grasping at straws here...