Add SpecifcValue and GateLength modules
SpecificValue lets you set a specific value
to send as CV voltage. It also displays the
voltage, frequency, period, note name and
cents offset corresponding to the value.

The values can be set via a knob, via
editable text fields, or via the 'input'.

So in addition to being usable as a controller,
it can also function as a test meter, displaying
voltage, freq, and note name of incoming signal.

GateLength turns triggers or gates into
gates of a provided length and includes
a display of the gate length. The gate length
time is also controllable via cv input.
(GateLength is kind of spartan at the moment)

GateLength was based on discussion at
https://www.facebook.com/groups/vcvrack/permalink/161960391130780/