Add a ValueSaver module, stores/remembers input value
Patch a signal though, and the CV value will be
saved.

Will remember the last value for an input that
has sense had it's input wire removed.

Will also continue to output the last input
value if the input is wired up, but it is
staying at 0. ie, like a Midi-1 with a CV
output that hasn't sent anything out yet.

If you use Midi controlers, this provides a
way to remember the CV value sent by the controller
across vcv rack open/close without having to
resend it from the device.