Add 'Hovered Value" module
This module shows the values of param widgets
(knobs, sliders, switched, etc) and input/output
ports as you hover over them in the gui.

The current value is displayed
as well as it's min, max, and default values.
And a 'type' to indicate if a
param/input/output.

The 'out' port will send out whatever value
the hovered port is currently set to.
Note: currently not clamped to -10/+10V